url error handling
